requirements.txt
Flask
flask-cors
langchain
langchain-chroma
langchain-openai
openai
chromadb
python-dotenv



pypdf
unstructured

